finally got the damn database to work
