+++ routes/mainController.js
@@ -1,7 +1,35 @@
 var express = require('express');
 var router = express.Router();
 
+var theDB = require('../util/db');
+
+var mongoose = require('mongoose');
+mongoose.connect('mongodb://appControl:control1@ds145704.mlab.com:45704/heroku_r2hv5571', {useNewUrlParser: true});
+var db = mongoose.connection;
+
+var employeeSchema = new mongoose.Schema({
+    name: String,
+    id: Number,
+    email: String,
+    uname: String,
+    pass: String,
+    groupIDs: Array,
+    monAvail: Array
+},{collection: 'employee'});
+
+var EmployeeModel = db.model('Employee', employeeSchema);
+
 //var mysql = require('mysql');
+
+db.once('open', function(){
+    console.log('Connected to database');
+    //test
+    //can comment the below statement out later, this proves that the db
+    //can be accessed and data can be pulled from it
+    theDB.getAll(EmployeeModel).then(function(doc){
+        console.log(JSON.stringify(doc));
+    });
+});
 
 var viewAddress;
 var viewData;
